requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
CodeforcesApiPy==1.3.1
dateparser==0.7.4
Flask==1.1.2
Frozen-Flask==0.15
icalendar==4.0.6
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.5.1
MarkupSafe==1.1.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
regex==2020.5.14
requests==2.23.0
six==1.14.0
tqdm==4.46.0
tzlocal==2.1
urllib3==1.25.9
Werkzeug==1.0.1
